refactor 비지니스 로직과 UI분리
gameController추가, RacingGame에서 메인메서드 분리
메인메서드 분리에 따른 RacingGame 일부 메서드 접근제어자 변경
